+++ dash_renderer/__init__.py
@@ -1,3 +1,5 @@
+import sys
+
 # For reasons that I don't fully understand,
 # unless I include __file__ in here, the packaged version
 # of this module will just be a .egg file, not a .egg folder.
@@ -10,9 +12,8 @@
 from .version import __version__
 __file__
 
-REACT_VERSION = '15.4.2'
+_DEFAULT_REACT_VERSION = '15.4.2'
 _REACT_VERSION_TYPES = {'15.4.2', '16.2.0'}
-assert REACT_VERSION in _REACT_VERSION_TYPES
 _REACT_VERSION_TO_URLS = {
     '15.4.2': {
         'external_url': [
@@ -36,15 +37,39 @@ _REACT_VERSION_TO_URLS = {
     }
 }
 
-# Dash renderer's dependencies get loaded in a special order by the server:
-# React bundles first, the renderer bundle at the very end.
-_js_dist_dependencies = [
-    {
-        'external_url': _REACT_VERSION_TO_URLS[REACT_VERSION]['external_url'],
-        'relative_package_path': _REACT_VERSION_TO_URLS[REACT_VERSION]['relative_package_path'],
+
+def _set_react_version(react_version):
+    """
+    Update the version of React in _js_dist_dependencies served by dash-renderer to the client
+
+    Example:
+    ```
+    import dash_renderer
+
+    # Set the react version before setting up the Dash application
+    dash_renderer._set_react_version('16.2.0')
+
+    app = dash.Dash(...)
+    ```
+
+    :param str react_version: Version of React
+
+    """
+    assert react_version in _REACT_VERSION_TYPES
+
+    _this_module = sys.modules[__name__]
+
+    # Dash renderer's dependencies get loaded in a special order by the server:
+    # React bundles first, the renderer bundle at the very end.
+    setattr(_this_module, '_js_dist_dependencies', [{
+        'external_url': _REACT_VERSION_TO_URLS[react_version]['external_url'],
+        'relative_package_path': _REACT_VERSION_TO_URLS[react_version]['relative_package_path'],
         'namespace': 'dash_renderer'
-    }
-]
+    }])
+
+
+_js_dist_dependencies = []
+_set_react_version(_DEFAULT_REACT_VERSION)
 
 _js_dist = [
     {